+++ posthog/temporal/batch_exports/squash_person_overrides.py
@@ -12,6 +12,7 @@ from temporalio.common import RetryPolicy
 
 from posthog.temporal.batch_exports.base import PostHogWorkflow
 from posthog.temporal.common.clickhouse import get_client
+from posthog.temporal.common.heartbeat import heartbeat_every
 
 EPOCH = datetime(1970, 1, 1, 0, 0, tzinfo=timezone.utc)
 
@@ -225,37 +226,6 @@ def parse_mutation_command(mutation_query: str) -> str:
         raise ValueError("Provided query does not appear to be an 'ALTER TABLE ... ON CLUSTER' mutation")
 
     return query_command
-
-
-def no_details() -> tuple:
-    """No heartbeat details."""
-    return ()
-
-
-@contextlib.asynccontextmanager
-async def heartbeat_every(
-    factor: int = 2,
-    details_callable: collections.abc.Callable[[], tuple[typing.Any]] = no_details,
-) -> collections.abc.AsyncIterator[None]:
-    """Heartbeat every Activity heartbeat timeout / factor seconds while in context."""
-    heartbeat_timeout = activity.info().heartbeat_timeout
-    heartbeat_task = None
-
-    async def heartbeat_forever(delay: float) -> None:
-        """Heartbeat forever every delay seconds."""
-        while True:
-            await asyncio.sleep(delay)
-            activity.heartbeat(*details_callable())
-
-    if heartbeat_timeout:
-        heartbeat_task = asyncio.create_task(heartbeat_forever(heartbeat_timeout.total_seconds() / factor))
-
-    try:
-        yield
-    finally:
-        if heartbeat_task:
-            heartbeat_task.cancel()
-            await asyncio.wait([heartbeat_task])
 
 
 @activity.defn
